Navigation added
Added navigation support with two example views:
Home and StepByStepList.

The current viewmodel is stored in NavigationStore
in it's _currentViewModel property.

The MainViewModel is the DataContext
extension view: ExtensionWindowControl.xaml.
Which uses the property to show the correct view.

Views can change the current viewmodel by changing
the store's property.
This is done through the Navigate Commands:
NavigateHomeCommand & NavigateStepByStepListCommand.
These commands are implemented in the ViewModels and used
in the views through button Bindings.