Rec: added typecheck to propertycheck and methodcheck
Did also add PropertyCheck to the options of ClassCheck
Did not write explicit tests to test this. Did however see whether it works, which it does.
I think it is not necessary because this should be tested for the recognizer runner, but if desired I would gladly write some extra tests.