Rec: Run Class/Interface checks on all entities
Class and Interface checks now iterate over all entities, and can now
also contain multiple matches like a MethodCheck. This enables
recognizers to check for multiple entities and relations between the
entities.

To determine whether a ClassCheck isn't incorrectly nested, the parent
check is stored on the RecognizerContext. This context is refactored in
such a way that it is impossible to create a new instance outside of
RecognizerRunner.

As the initial root check doesn't have a CurrentEntity yet, and we don't
want to deal with that property being null everywhere, the RootNode
class is introduced which serves as a kind of sentinel value for the
root check.