SBS: Use old way of determining if result is correct
Due to some limitations in the score calculation, we unfortunately can't
rely on that to determine if the result is correct. Instead we fallback
to the old way of determining that, which is to run with prune all and
check if any results remain. This does mean that the recognizer
technically runs twice, which is inefficient. However, because a
Step-by-Step implementation is usually quite small, this doesn't really
matter.