Feat: SBS continue opens more convenient initial directory in the dialog. (#390)
Currently, the file dialog for continue has an initial directory, the
last remembered directory. This PR changes it so that it looks for
projects in the opened solution. If the name of that project corresponds
to the selected design pattern from the drop-down bar it uses that
project's folder path as an initial directory. If there is no such
project it opens the solution folder path instead that is loaded and if
there is neither a project or a solution it has the same behaviour as
currently.